refactor: extract Provider.Staff, reduce facade to pure delegation
Move all staff-member commands, invitation lifecycle, reads, the context
switcher, and their helpers onto Provider.Staff; emit_staff_invitation now
calls Profiles.get_provider_profile/1 directly. The KlassHero.Provider facade
is now a thin, logic-free defdelegate layer over the six sub-domain modules
(1265 -> 212 LOC).

Closes #1003
